chore: separate scene segmentation handler from volume handler
